% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goto.R
\name{goto}
\alias{goto}
\title{goto}
\usage{
goto(label)
}
\arguments{
\item{label}{label of the goto}
}
\value{
called for side effects
}
\description{
call \code{goto(label)} and it'll move you to \code{(label)}
}
\examples{
foo <- function() {
  print("a")
  goto("my label")
  print("b")
  ("my label")
  print("c")
}
foo()
}
